vector-store: add Filter and Restriction types
The commit adds Restriction and Filter types to the vector-store. Filter is a
collector for restrictions which will be used to filter results in usearch.

The commit adds FilteredAnn to the index actor to allow search with filter in
the index.

The specific filtering in usearch will be introduced in next commits. The
filtered search will be run in rayon thread as it could be much slower than
search.