vector-store: add table columns to db_index
To support filtering we need to access the map between table column name and
its cql type. A cql type will be needed for deserializing json value to the
CqlValue. The commit introduces new message to the db_index actor:
GetTableColumns.

Finally, the commits adds support for GetTableColumns in integration tests and
db_mock.